+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Option to configure reservation pods runtime class.
+
 ## [v0.9.5] - 20250-10-09
 
 ### Added